split code into Scintilla_iface
