+++ Visual Studio Project Template C#/$projectname$.csproj
@@ -52,6 +52,7 @@
     <Compile Include="NppPluginNETHelper.cs" />
     <Compile Include="Integration\Docking_h.cs" />
     <Compile Include="Integration\MenuCmdID_h.cs" />
+    <Compile Include="Integration\Scintilla_iface.cs" />
     <Compile Include="Integration\Msgs_h.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
